Recent updates and small bugfixes
Found a few odds and ends in the front and back that I cleaned up for a few UX fixes. Instead of remaking a new quick options page, I went with Michaels and just added in some text and verified redirects. Ran into some issues with how information is called to the "re-order my favorite" section though so if we can collab on this that would be great!