Update to Surprise Me
"Surprise me" just led to the same craft_a_pizza page. In order to have it populate something random though, I felt the best way was to have each button on quick options lead to a different instance of the craft_a_pizza. I made a new app.route to do this with a dictionary of each portion of the pizza's attributes with a random function. This has partially worked, however the next step is to find a way to just grab the first value and not both the attribute value, price, and key altogether.